Make sure the result is deterministic if 2 candidates have the same content_score.
If the 2 top candidates had the same content_score, the 1 selected as the best candidate in select_best_candidate would be non-deterministic. It we determined by the order candidates.values() which seems to be based on the memory address of elements. Thus, if the same page is processed multiple times, readability could return different results.

To fix this I have added a candidate_number field whose value is determined by the processing order in score_paragraphs.